Fix parsing and changes
The most relevant change is that for the new challenge a new email is used.
The ACME accounts are bound to the URL (including IP:PORT). So when changing
the endpoint to a separate port number the account is missing. LEGO finds the
account based on the email and then there is also a key mismatch (if transiting
to PQC). Using a new email forces a new account creation in an easier manner.
Other fixes:

* Certificate reading from path was null, because readResource reads only json.
Now it reads based on the extension (.crt and .key).
* Change messages in the logs
* Fix request type (application/jose+json)

Note: InsecureSkipVerify is being used for testing the request messages.
Should change this later on. Pebble replies certificate is issued
but LEGO is not handling the reply codes correctly. So it's a WiP again.